refactor(policy): extract declaration machinery into mureo/policy/declarations.py
Pure move, zero behavior change: BudgetDeclaration, BidDeclaration,
their registries and register/lookup/reset helpers, and the shared
declared-key readers (_Unreadable, _saturate, _declared_amount) move to
a new sibling module, bringing strategy_gate.py from 877 back under the
800-line budget (714 lines; declarations.py is 244).

strategy_gate re-exports every moved name (grouped import + __all__,
including the three underscore names the test suite imports) so the
existing import path remains a stable ABI for the bridges; the shared
registry dicts stay identical objects across both module paths.
Internal callers (plugin_semantics, server registration) repoint at the
canonical module. Existing gate and declaration test files unmodified.
